Supports better algorithm for removing by key
Fixed a bug where slice was used instead of concat for producing an
array of elements. I explicitly filter elements to detached/attached to
be real elements (no text nodes).

This includes a smarter remove for key attributes. In the future it'd be
nice to do a replace by key as well for sorting.