Taking the time to clean up the codebase
I've wanted to take the time to clean up the codebase and ensure
everything is consistent. Right now my refactor efforts are to use
common nomenclature when talking about Virtual Tree Elements and DOM
Nodes. This distinction is important in the codebase, and is currently
different depending on the file you're in. Sometimes they're called
entry, sometimes tree, descriptor, etc. I'm consolidating all the
variable names to vTree and node.

Another important refactor is to provide clean and consistent JSDoc
comments above all functions public/private.

I'm attempting to use `const` wherever possible to denote that a
variable is never re-assigned. I'm refactoring branched code that
required using `var` or `let` into functions. I will still use `let` for
loops, since we are tight-loop critical and I don't like the overhead of
functional looping here.

There are parts of the codebase that are highly duplicative and I'm
leveraging the createElement/createAttribute functions, exposed for the
Babel transform, internally.

In preparation for a 1.0 release, I want to make sure all the code makes
sense and references variables consistently. This furthers that push by
making it explicit that an object is either a DOM Node or a Virtual Tree
Element/Attribute. This made it easier to re-arrange the file structure
to better match what they do. For instance, before makeElement made a
DOM Node and makeNode made a Virtual Tree Element.

This didn't affect tests other than for import paths in the internal
unit tests.