Rewrite the synchronization and patching methods
Removing the MODIFY_ELEMENT operation in favor of APPEND, PREPEND,
REPLACE, and REMOVE. The MODIFY_ATTRIBUTE is so simple and accounts for
all cases, so that will most likely stay.

Changing the patch from __do__ to type which aligns better with Redux
actions.

Having a baton that carries the patches and current operation allowing
for a more standardized approach to applying patches. If multiple of the
same type of patch occurs, they can be batched into DocumentFragments.
I'm not sure if many developers know this, but removeChild and
insertBefore both support Fragments.

Remove REMOVE_ENTIRE_ELEMENT, REPLACE_ENTIRE_ELEMENT, and
REMOVE_ELEMENT_CHILDREN in favor of the new aforementioned element
operation types.

Miscellaneous changes:

- Always ensure valid state names are passed to removeTransitionState
- Refactor finalize transaction
- A lot more tests and refactoring around tree synchronization