Changes to project structure, consistency updates
diffHTML has been a radical collection of good ideas poorly implemented.
This commit attempts to reverse some of the bad code ideas, and make the
project generally more consistent. I'm starting with `createTree`,
before this was known as `createElement` and was a way of create
virtual trees. This new method has a better name and role, where-as it
is called on any tagged template child value automatically and will
simply be an identity function for any VTree passed to it. This ensures
that all trees are fully formed before a transaction ever starts.

I'm going to continue the refactor by gutting and redoing the tree
synchronization code. I'm going to focus on making patches that are
driven by the DOM method ultimately called. This will make applying the
patches incredibly easy and allow us to batch attributes from DOM
methods, which was not always the case before.