Lowercases all event names, normalizes to 'on*'
This commit updates the README to flatten all event names to lower case,
but does not require it. It updates the signature of the event lookup to
be: 'onattached' from 'attached', this indicates its more of an event
that is frequently called. Since none of these events currently
collide with browser events it shouldn't be an issue. Although if they
ever do, I'm sure this project may be long dead by then...