Big commit. Like HUUUUGE. WC & ES & CJS oh my...
This adds in the big kahuna burger of direction. Has tests passing for
React Like Components, Web Components, a shared internals module, better
ES and CJS output for diffhtml, diffhtml-components, and
diffhtml-shared-internals to start with. Will roll out to the rest if
this works out well.

I've been experimenting with JSDOM and WC integration and released a
module JSDOM-WC that supports enough to get testing going. Hopefully it
will get merged into proper.