Start reworking component implementation
The current implementation is broken and doesn't work consistently which
results in broken UI. This is an initial attempt to rework how stateful
components are re-rendered. The work isn't complete and this is not yet
ready.

If you are interested in component architecture and systems and want to
lend a hand it would be greatly appreciated. For the most part this
functions correctly, but there are lots of edge cases that still need to
be ironed out.

Since this is of upmost importance to shipping a release candidate,
moreso than any other remaining work, I will be focusing on this until
it is completed.